Organize async resource loading
