+++ Spatial.Core/CMakeLists.txt
@@ -79,6 +79,7 @@ add_library(Spatial.Core
         include/spatial/resources/ResourceError.h
         include/spatial/resources/FileSystem.h
         include/spatial/resources/PhysicalFileSystem.h
+        include/spatial/resources/ResourceController.h
 
         include/spatial/script/Isolate.h
         include/spatial/script/PlatformContext.h
@@ -138,7 +139,8 @@ add_library(Spatial.Core
         src/resources/FilameshFile.cpp
         src/resources/FileSystem.cpp
         src/resources/PhysicalFileSystem.cpp
-        src/resources/MemoryFileSystem.cpp include/spatial/ecs/ResourceController.h src/ecs/ResourceController.cpp)
+        src/resources/MemoryFileSystem.cpp
+        src/resources/ResourceController.cpp)
 
 set_project_warnings(Spatial.Core)
 enable_sanitizers(Spatial.Core)
