+++ Spatial.Core/include/spatial/ecs/Resource.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <spatial/ecs/Registry.h>
+#include <vector>
 
 namespace spatial::ecs
 {
