+++ Spatial.Core/src/resources/ResourceController.cpp
@@ -2,12 +2,10 @@
 #include <future>
 #include <spatial/core/Logger.h>
 #include <spatial/ecs/Resource.h>
-#include <spatial/ecs/ResourceController.h>
+#include <spatial/resources/ResourceController.h>
 
-namespace spatial::ecs
+namespace spatial
 {
-
-static auto gLogger = createDefaultLogger();
 
 void ResourceController::loadResources(FileSystem& fileSystem, ecs::Registry& registry)
 {
@@ -15,7 +13,7 @@ void ResourceController::loadResources(FileSystem& fileSystem, ecs::Registry& re
 	using FutureData = std::future<std::vector<uint8_t>>;
 
 	registry
-		.getEntities<const ecs::Resource>(ecs::ExcludeComponents<FutureData, ResourceData, ecs::tags::IsResourceLoaded>)
+		.getEntities<const ecs::Resource>(ecs::ExcludeComponents<FutureData, ecs::ResourceData, ecs::tags::IsResourceLoaded>)
 		.each([&](ecs::Entity entity, const auto& resource) {
 			if (resource.relativePath.empty())
 				return;
@@ -24,16 +22,15 @@ void ResourceController::loadResources(FileSystem& fileSystem, ecs::Registry& re
 			registry.addComponent(entity, std::move(futureData));
 		});
 
-	registry
-		.getEntities<const ecs::Resource, FutureData>(ecs::ExcludeComponents<ResourceData>)
+	registry.getEntities<const ecs::Resource, FutureData>(ecs::ExcludeComponents<ecs::ResourceData>)
 		.each([&](ecs::Entity entity, const auto& resource, auto& futureData) {
 			if (futureData.wait_for(10us) != std::future_status::ready)
 				return;
 
 			auto data = futureData.get();
-			registry.addComponent<ResourceData>(entity, std::move(data));
+			registry.addComponent<ecs::ResourceData>(entity, std::move(data));
 			registry.removeComponent<FutureData>(entity);
 		});
 }
 
-} // namespace spatial::ecs
+} // namespace spatial
